requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-rest-auth==0.9.3
djangorestframework-jwt==1.11.0
Pillow==5.2.0
django-cors-headers==3.2.0